% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_find.R
\name{article_find}
\alias{article_find}
\title{Extracts all urls on journal sitemap but keeps only papers. Outputs a vector with urls}
\usage{
article_find(journal)
}
\arguments{
\item{journal}{A string containing the name of a MDPI journal}
}
\value{
A vector
}
\description{
Extracts all urls on journal sitemap but keeps only papers. Outputs a vector with urls
}
\examples{
article_find("agriculture")
}
