% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_info.R
\name{article_info}
\alias{article_info}
\title{Extracts data for each individual paper}
\usage{
article_info(vector, sleep = 2, sample_size)
}
\arguments{
\item{vector}{A vector with urls.}

\item{sleep}{Number of seconds between scraping iterations. 2 sec. by default}

\item{sample_size}{A number. How many papers do you want to explore from the main vector. Leave blank for all}
}
\value{
A data frame.
}
\description{
Extracts data for each individual paper
}
\examples{
url<-c("https://www.mdpi.com/2073-4336/8/4/45","https://www.mdpi.com/2073-4336/11/3/39")
\dontrun{
articles_info(url, 1.5)
}
}
