% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_graph.R
\name{average_graph}
\alias{average_graph}
\title{Creates a plot with monthly turnaround times using the obtained via articles_info()}
\usage{
average_graph(articles_info, journal)
}
\arguments{
\item{articles_info}{Output dataframe from function articles_info.}

\item{journal}{A string with the name of the journal for graph title purposes}
}
\value{
A plot
}
\description{
Creates a plot with monthly turnaround times using the obtained via articles_info()
}
\examples{
average_graph(agriculture,"Agriculture")
}
