% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guest_editor.R
\name{guest_editor}
\alias{guest_editor}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Function was deprecated and split in special_issue_finder and guest_editor_info()}
\usage{
guest_editor(journal, sleep)
}
\arguments{
\item{journal}{MDPI journal code}

\item{sleep}{Number of seconds between scraping iterations}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Function was deprecated and split in special_issue_finder and guest_editor_info()
}
\examples{
\dontrun{
guest_editor("covid", 1.5)
}
}
