% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guest_editor_info.R
\name{guest_editor_info}
\alias{guest_editor_info}
\title{Finds data on papers published by guest editors in their own special issues}
\usage{
guest_editor_info(journal_urls, sample_size, sleep = 2)
}
\arguments{
\item{journal_urls}{MDPI journal code}

\item{sample_size}{A number. How many special issues do you want to explore from the main vector. Leave blank for all}

\item{sleep}{Number of seconds between scraping iterations. 2 sec. by default}
}
\value{
A data frame.
}
\description{
Finds data on papers published by guest editors in their own special issues
}
\examples{
\dontrun{
guest_editor_info("covid", 1.5)
}
}
