% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horticulturae.R
\docType{data}
\name{horticulturae}
\alias{horticulturae}
\title{Article data extracted from MDPI journal Horticulturae}
\format{
\subsection{\code{horticulturae}}{

A data frame with 7,160 rows and 7 columns:
\describe{
\item{i}{Article URL}
\item{article_type}{Article tyope classifier}
\item{Received}{Date article was submitted to journal}
\item{Accepted}{Date article was accepted for publication}
\item{tat}{Article turnaround time, or Accepted-Received}
\item{year}{Year the article was accepted}
\item{issue_type}{Type of issue where article is published}
...
}
}
}
\usage{
horticulturae
}
\description{
Article data extracted from MDPI journal Horticulturae
}
\keyword{datasets}
