% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_graph.R
\name{issues_graph}
\alias{issues_graph}
\title{Creates a plot showing types of issues were articles where published}
\usage{
issues_graph(articles_info, journal)
}
\arguments{
\item{articles_info}{Output dataframe from function articles_info.}

\item{journal}{A string with the name of the journal for graph title purposes}
}
\value{
A plot
}
\description{
Creates a plot showing types of issues were articles where published
}
\examples{
issues_graph(agriculture,"Agriculture")
}
