% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_papers.R
\name{journal_papers}
\alias{journal_papers}
\title{Extracts all urls on journal sitemap but keeps only papers. Outputs a vector with urls}
\usage{
journal_papers(journal)
}
\arguments{
\item{journal}{A string containing the name of a MDPI journal}
}
\value{
A vector
}
\description{
Extracts all urls on journal sitemap but keeps only papers. Outputs a vector with urls
}
\examples{
journal_papers("agriculture")
}
