% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDPI_journals.R
\docType{data}
\name{MDPI_journals}
\alias{MDPI_journals}
\title{List of 430 MDPI journals and sitemap reference code}
\format{
## `MDPI_journals`
A data frame with 430 rows and 2 columns:
\describe{
  \item{name}{Journal name}
  \item{code}{Journal code used in the sitemap}
  ...
}
}
\usage{
MDPI_journals
}
\description{
List of 430 MDPI journals and sitemap reference code
}
\keyword{datasets}
