% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfcite_check.R
\name{selfcite_check}
\alias{selfcite_check}
\title{Calculates number of authors selfcitations against all references}
\usage{
selfcite_check(article_url)
}
\arguments{
\item{article_url}{A valid MDPI article url}
}
\value{
A string.
}
\description{
Calculates number of authors selfcitations against all references
}
\examples{
\dontrun{
selfcite_check(paper_url)
}
}
