% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graph.R
\name{summary_graph}
\alias{summary_graph}
\title{Creates a summary of the dataframe obtained via articles_info()}
\usage{
summary_graph(articles_info, journal)
}
\arguments{
\item{articles_info}{Output dataframe from function articles_info.}

\item{journal}{A string with the name of the journal for graph title purposes}
}
\value{
A plot
}
\description{
Creates a summary of the dataframe obtained via articles_info()
}
\examples{
summary_graph(agronomy,"Agronomy")
}
