% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_find.R
\name{article_find}
\alias{article_find}
\title{This function retrieves the URLs for all published articles from a specified journal. Users can provide the journal's code 'see MDPI_journals.rda', and the function will return the URLs of all articles available within the journal.}
\usage{
article_find(journal)
}
\arguments{
\item{journal}{A string containing the name of a MDPI journal}
}
\value{
A vector (class: \code{character}) containing a list of articles URLs from target journal
}
\description{
This function retrieves the URLs for all published articles from a specified journal. Users can provide the journal's code 'see MDPI_journals.rda', and the function will return the URLs of all articles available within the journal.
}
\examples{
\dontrun{
agr_articles<-article_find("agriculture")
}
}
