% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_footprint.R
\name{guest_editor}
\alias{guest_editor}
\title{Finds data on papers published by guest editors in their own special issues}
\usage{
guest_editor(journal, sleep)
}
\arguments{
\item{journal}{MDPI journal code}

\item{sleep}{Number of seconds between scraping iterations}
}
\value{
A data frame.
}
\description{
Finds data on papers published by guest editors in their own special issues
}
\examples{
\dontrun{
guest_editor("covid", 1.5)
}
}
