% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guest_editor_info.R
\name{guest_editor_info}
\alias{guest_editor_info}
\title{#' Obtain information from guest edited special issues}
\usage{
guest_editor_info(journal_urls, sample_size, sleep = 2)
}
\arguments{
\item{journal_urls}{A list of MDPI special issues URLs}

\item{sample_size}{A number. How many special issues do you want to explore from the main vector. Leave blank for all}

\item{sleep}{Number of seconds between scraping iterations. 2 sec. by default}
}
\value{
A data frame.
}
\description{
#'
Extracts data from special issues, including guest editors' paper counts excluding editorials, time between last submission and issue closure, and whether guest editors served as academic editors for any published papers.
}
\examples{
\dontrun{
guest_editor_info("covid", 1.5)
}
}
