% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDPI_journals.R
\name{MDPI_journals}
\alias{MDPI_journals}
\title{MDPI journal names and code}
\usage{
MDPI_journals()
}
\value{
A data frame (class: \code{data.frame}) with the following columns:
\describe{
\item{journal}{Full name of the MDPI journal}
\item{num_papers}{Journal code used for ID and web scraping purposes}
}
}
\description{
Extracts names and codes of current MDPI journals.
}
\examples{
\dontrun{
journal_table<-MDPI_journals()
}

}
