% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_articles.R
\name{plot_articles}
\alias{plot_articles}
\title{Plots information obtained from article_info(). For analysis purposes, Editorial and Correction type articles are ignored.}
\usage{
plot_articles(articles_info, journal, type)
}
\arguments{
\item{articles_info}{Output dataframe from function articles_info.}

\item{journal}{A string with the name of the journal for graph title purposes}

\item{type}{select "summary","issues", "tat", "review" or "type" depending on desired graph}
}
\value{
A plot (class: \code{ggplot}) depicting the desired information obtained from \code{article_info}
}
\description{
Plots information obtained from article_info(). For analysis purposes, Editorial and Correction type articles are ignored.
}
\examples{
plot_articles(agriculture,"Agriculture",type="summary")

}
