% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfcite_check.R
\name{selfcite_check}
\alias{selfcite_check}
\title{Calculates number of authors selfcitations against all references}
\usage{
selfcite_check(article_url)
}
\arguments{
\item{article_url}{A valid MDPI article url}
}
\value{
A string (class: \code{character}) containg the values of possible self citations vs total number of references in the paper
}
\description{
Calculates number of authors selfcitations against all references
}
\examples{
paper_url<-"https://www.mdpi.com/2223-7747/13/19/2785"
sc<-selfcite_check(paper_url)

}
