% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{types_graph}
\alias{types_graph}
\title{Creates a plot showing a summary of article types}
\usage{
types_graph(article_info, journal)
}
\arguments{
\item{journal}{A string with the name of the journal for graph title purposes}

\item{articles_info}{Output dataframe from function articles_info.}
}
\value{
A plot
}
\description{
Creates a plot showing a summary of article types
}
\examples{
types_graph(agronomy,"Agronomy")
}
